1. Fixed issue where the current site could be removed if too many broken backups were created after the symlinked one. 2. Fixed an issue where on some systems the sinatra-based site wouldn't work.

+++ lib/git.rb
@@ -25,77 +25,78 @@ module Ripple
 
     def clone
       Dir.mkdir @backup_dir
-      Dir.chdir @backup_dir
 
-      io_log 'git init'
-      io_log "git remote add origin #{ @opts[:repo] }"
-
-      Dir.chdir @backup_dir
+      Dir.chdir @backup_dir do
+        io_log 'git init'
+        io_log "git remote add origin #{ @opts[:repo] }"
+      end
 
       post_process @backup_dir
     end
 
     def process
       backups = Dir.glob(@backup_prefix + '*')
+      path    = (File.symlink? @directory) ? File.readlink(@directory) : @directory
 
       if backups.length > @opts[:backups]
         backups.sort!
         extra = backups.length - @opts[:backups]
         backups.slice(0...extra).each do |b|
+          FileUtils.rm_r b unless path == b
           msg = "Deleted: #{ b }"
+
           puts msg if @console
           @log << msg + "\n"
-
-          FileUtils.rm_r b
         end
       end
 
-      path = (File.symlink? @directory) ? File.readlink(@directory) : @directory
-
       FileUtils.rm_r @backup_dir if Dir.exists? @backup_dir
       FileUtils.cp_r path, @backup_dir
+      File.delete @backup_dir + '/deployment.log' if File.exists? @backup_dir + '/deployment.log'
 
-      Dir.chdir @backup_dir
-      io_log 'git reset --hard HEAD'
+      Dir.chdir @backup_dir do
+        io_log 'git reset --hard HEAD'
+      end
 
       post_process @backup_dir
     end
 
     def post_process(temp_dir)
-      Dir.chdir temp_dir
-      io_log "git pull #{ @opts[:git][:remote] } #{ @opts[:git][:branch] }"
+      Dir.chdir temp_dir do
+        io_log "git pull #{ @opts[:git][:remote] } #{ @opts[:git][:branch] }"
 
-      # For PHP projects using composer
-      if File.exists? 'composer.json'
-        composer = @opts[:composer].has_key?(:path) ? @opts[:composer][:path] : temp_dir + '/composer.phar'
-        unless File.exists? composer
-          c = Net::HTTP.get_response(URI.parse('http://getcomposer.org/installer')).body
-          File.write(composer, c)
+        # For PHP projects using composer
+        if File.exists? 'composer.json'
+          composer = @opts[:composer].has_key?(:path) ? @opts[:composer][:path] : temp_dir + '/composer.phar'
+          unless File.exists? composer
+            c = Net::HTTP.get_response(URI.parse('http://getcomposer.org/installer')).body
+            File.write(composer, c)
+          end
+
+          FileUtils.rm_r 'vendor' if @opts[:composer][:vendor] == 'clean' && Dir.exists?('vendor')
+
+          io_log "php #{ composer } self-update" if FileTest.writable? File.dirname(composer)
+          io_log "php #{ composer } #{ @opts[:composer][:command] } --prefer-#{ @opts[:composer][:source] } #{ @opts[:composer][:flags] }"
         end
 
-        FileUtils.rm_r 'vendor' if @opts[:composer][:vendor] == 'clean' && Dir.exists?('vendor')
+        # For Ruby projects using bundler
+        if File.exists? 'Gemfile'
+          # Do some bundler stuff
+          io_log 'sudo bundle install'
+        end
 
-        io_log "php #{ composer } self-update" if FileTest.writable? File.dirname(composer)
-        io_log "php #{ composer } #{ @opts[:composer][:command] } --prefer-#{ @opts[:composer][:source] } #{ @opts[:composer][:flags] }"
-      end
+        # Post executable files
+        if @opts.has_key? :post_exec
+          msg = 'Running post_exec scripts...'
+          puts msg if @console
+          @log << msg + "\n"
 
-      # For Ruby projects using bundler
-      if File.exists? 'Gemfile'
-        # Do some bundler stuff
-        io_log 'sudo bundle install'
-      end
+          @opts[:post_exec].each do |e|
+            puts e if @console
+            @log << e + "\n"
 
-      # Post executable files
-      if @opts.has_key? :post_exec
-        msg = 'Running post_exec scripts...'
-        puts msg if @console
-        @log << msg + "\n"
-
-        @opts[:post_exec].each do |e|
-          puts e if @console
-          @log << e + "\n"
-
-          io_log e
+            io_log e
+          end
         end
       end
 
@@ -108,12 +109,13 @@ module Ripple
         puts msg if @console
         @log << msg + "\n"
 
-        Dir.chdir @directory
-        @opts[:final_exec].each do |e|
-          puts e if @console
-          @log << e + "\n"
+        Dir.chdir @directory do
+          @opts[:final_exec].each do |e|
+            puts e if @console
+            @log << e + "\n"
 
-          io_log e
+            io_log e
+          end
         end
       end
